Create puzzle-game.py
In this example, we're defining a list of words for the puzzle using the words variable. We're then choosing a random word from the list using the random.choice() function.

We're creating a list named guesses to store the letters that the player has guessed. We're then looping until the player solves the puzzle or runs out of guesses.

Inside the loop, we're displaying the current state of the puzzle using a list comprehension. If a letter has been guessed, we're displaying it. Otherwise, we're displaying an underscore character.

We're then getting a guess from the player using the input() function. We're checking if the guess is correct by seeing if it's in the word variable. If the guess is correct, we're appending it to the guesses list. If the guess is incorrect, we're printing a message to tell the player.

We're then checking if the player has won or lost the game. If the set of the word variable and the guesses list are equal, then the player has guessed all the correct letters and has won the game. If the player has guessed six incorrect letters, then they have lost the game.